PR Revision - moved S3 methods to S3IT class
